Move identifier package to public api instead of internal
